requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
dacite==1.9.2
frozenlist==1.7.0
fuzzywuzzy==0.18.0
idna==3.10
Levenshtein==0.27.1
multidict==6.6.3
propcache==0.3.2
python-dotenv==1.1.1
python-Levenshtein==0.27.1
RapidFuzz==3.13.0
requests==2.32.4
themoviedb==1.0.2
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
